moved the types to extra folder
